Update bot functionality and add new commands
